moved neo4j tests to "integrations tests"
